packages/be/requirements.txt
fastapi[standard]
starlette
python-socketio